chore(api): move validator logic to Validatable class to allow for non-entity validation
