+++ libs/api/shared/src/index.ts
@@ -1,6 +1,7 @@
 export * from './lib/models/request.model';
 export * from './lib/models/base-entity.model';
 export * from './lib/models/base-document.model';
+export * from './lib/models/validatable.model';
 export * from './lib/kernel/graphql';
 export * from './lib/kernel/mongodb';
 export * from './lib/kernel/shared-kernel.module';
